backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Compress==1.15
Flask-Limiter==3.8.0
python-dotenv==1.0.1
google-generativeai==0.8.3
gunicorn==21.2.0
